reduce compoent layers
