+++ packages/components/src/components/Controls/index.js
@@ -1,14 +1,9 @@
 import BaseControl from './BaseControl';
-import CheckboxControl from './CheckboxControl';
-import RadioControl from './RadioControl';
-import TextControl from './TextControl';
+
 import TextAreaControl from './TextAreaControl';
 import ToggleControl from './ToggleControl';
 export {
-	CheckboxControl,
 	BaseControl,
-	RadioControl,
-	TextControl,
 	TextAreaControl,
 	ToggleControl
 }
