+++ packages/components/src/components/fields/RadioField/RadioField.js
@@ -1,7 +1,9 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import { parseAttributes, fieldClassNames, isValidHtml5type } from '../util';
-import RadioControl from "../../Controls/RadioControl";
+import {isEmpty} from "lodash";
+import BaseControl from "../../Controls/BaseControl";
+import classnames from "classnames";
 export const RadioField = ({
 	attributes,
 	label,
@@ -11,21 +13,36 @@ export const RadioField = ({
 	options,
 	multiple,
 	description,
-	placeholder
+	placeholder,
+							   className
 }) => {
 	attributes = parseAttributes(attributes, 'radio');
+	const onChangeValue = (event) => onChange(event.target.value);
 
-	return (
-		<RadioControl
-			className={fieldClassNames('radio')}
-			id={fieldId}
-			selected={value}
-			options={options}
-			onChange={onChange}
-			label={label}
-			help={description}
-			{...attributes}
-		/>
+
+	return !isEmpty(options) && (
+		<BaseControl label={label} id={fieldId} help={description} className={classnames(className, 'components-radio-control')}>
+			{options.map((option, index) =>
+				<div
+					key={`${ fieldId }-${ index }`}
+					className="components-radio-control__option"
+				>
+					<input
+						id={`${ fieldId }-${ index }`}
+						className="components-radio-control__input"
+						type="radio"
+						name={fieldId}
+						value={option.value}
+						onChange={onChangeValue}
+						checked={option.value === value}
+						aria-describedby={!!description ? `${ fieldId }__help` : undefined}
+					/>
+					<label htmlFor={`${ fieldId }-${ index }`}>
+						{option.label}
+					</label>
+				</div>
+			)}
+		</BaseControl>
 	);
 };
 
@@ -34,7 +51,8 @@ RadioField.propTypes = {
 	description: PropTypes.string,
 	fieldId: PropTypes.string,
 	required: PropTypes.bool,
-	multiple: PropTypes.bool
+	multiple: PropTypes.bool,
+	options: PropTypes.array
 };
 
 RadioField.defaultProps = {
