+++ packages/components/src/components/fields/SelectField/SelectField.js
@@ -1,7 +1,8 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import { parseAttributes, fieldClassNames, isValidHtml5type } from '../util';
-import SelectControl from "../../Controls/SelectControl";
+import {isEmpty} from "lodash";
+import BaseControl from "../../Controls/BaseControl";
 export const SelectField = ({
 	attributes,
 	label,
@@ -20,18 +21,37 @@ export const SelectField = ({
 			label: placeholder
 		});
 	}
-	return (
-		<SelectControl
-			className={fieldClassNames('select')}
-			id={fieldId}
-			value={value}
-			options={options}
-			onChange={onChange}
-			label={label}
-			help={description}
-			{...attributes}
-		/>
+
+	const onChangeValue = (event) => {
+		onChange(event.target.value);
+	};
+
+	// Disable reason: A select with an onchange throws a warning
+
+	/* eslint-disable jsx-a11y/no-onchange */
+	return !isEmpty(options) && (
+		<BaseControl label={label} id={fieldId} help={description} className={''}>
+			<select
+				id={fieldId}
+				className="components-select-control__input"
+				onChange={onChangeValue}
+				aria-describedby={!!description ? `${ fieldId }__help` : undefined}
+				multiple={multiple}
+				{...attributes}
+			>
+				{options.map((option, index) =>
+					<option
+						key={`${ option.label }-${ option.value }-${ index }`}
+						value={option.value}
+					>
+						{option.label}
+					</option>
+				)}
+			</select>
+		</BaseControl>
 	);
+	/* eslint-enable jsx-a11y/no-onchange */
+
 };
 
 SelectField.propTypes = {
